+++ android/quest/src/test/java/org/smartregister/fhircore/quest/ui/patient/details/QuestPatientTestResultScreenTest.kt
@@ -61,7 +61,8 @@ class QuestPatientTestResultScreenTest : RobolectricTest() {
       spyk(
         QuestPatientDetailViewModel(
           patientRepository = patientRepository,
-          patientItemMapper = patientItemMapper
+          patientItemMapper = patientItemMapper,
+          mockk()
         )
       )
     questPatientDetailViewModel.getDemographics(patientId)